+++ packages/graph-editor/src/components/flow/nodes/series/arithmeticNode.tsx
@@ -39,21 +39,6 @@ const ArithmeticNode = (props) => {
   return (
     <Stack direction="row" gap={4}>
       <HandleContainer type="target">
-        <Handle id="base">
-          <Stack direction="row" justify="between" gap={3} align="center">
-            <Label>Base</Label>
-            {input?.base ? (
-              <PreviewNumber value={input.base} />
-            ) : (
-              <TextInput
-                data-key="base"
-                value={state.base}
-                onChange={onChange}
-              />
-            )}
-          </Stack>
-        </Handle>
-
         <Handle id="stepsDown">
           <LabelNoWrap>Steps Down</LabelNoWrap>
           {input?.stepsDown ? (
@@ -80,6 +65,21 @@ const ArithmeticNode = (props) => {
           )}
         </Handle>
 
+        <Handle id="base">
+          <Stack direction="row" justify="between" gap={3} align="center">
+            <Label>Base</Label>
+            {input?.base ? (
+              <PreviewNumber value={input.base} />
+            ) : (
+              <TextInput
+                data-key="base"
+                value={state.base}
+                onChange={onChange}
+              />
+            )}
+          </Stack>
+        </Handle>
+
         <Handle id="increment">
           <Stack direction="row" justify="between" gap={3} align="center">
             <Label>Increment</Label>
@@ -89,6 +89,21 @@ const ArithmeticNode = (props) => {
               <TextInput
                 data-key="increment"
                 value={state.increment}
+                onChange={onChange}
+              />
+            )}
+          </Stack>
+        </Handle>
+
+        <Handle id="precision">
+          <Stack direction="row" justify="between" align="center" gap={3}>
+            <Text>Precision</Text>
+            {input.precision !== undefined ? (
+              <Text>{input.precision}</Text>
+            ) : (
+              <TextInput
+                data-key="precision"
+                value={state.precision}
                 onChange={onChange}
               />
             )}