+++ packages/graph-editor/src/components/flow/nodes/series/geometricNode.tsx
@@ -39,26 +39,6 @@ const GeometricNode = (props) => {
   return (
     <Stack direction="row" gap={4}>
       <HandleContainer type="target">
-        <Handle id="base">
-          <LabelNoWrap>Base</LabelNoWrap>
-          {input.base ? (
-            <PreviewNumber value={input.base} />
-          ) : (
-            <TextInput data-key="base" value={state.base} onChange={onChange} />
-          )}
-        </Handle>
-        <Handle id="ratio">
-          <LabelNoWrap>Ratio</LabelNoWrap>
-          {input.ratio ? (
-            <PreviewNumber value={input.ratio} />
-          ) : (
-            <TextInput
-              data-key="ratio"
-              value={state.ratio}
-              onChange={onChange}
-            />
-          )}
-        </Handle>
         <Handle id="stepsDown">
           <LabelNoWrap>Steps Down</LabelNoWrap>
           {input.stepsDown ? (
@@ -71,8 +51,9 @@ const GeometricNode = (props) => {
             />
           )}
         </Handle>
+
         <Handle id="steps">
-          <LabelNoWrap>Steps</LabelNoWrap>
+          <LabelNoWrap>Steps Up</LabelNoWrap>
           {input.steps ? (
             <PreviewNumber value={input.steps} />
           ) : (
@@ -82,6 +63,43 @@ const GeometricNode = (props) => {
               onChange={onChange}
             />
           )}
+        </Handle>
+
+        <Handle id="base">
+          <LabelNoWrap>Base</LabelNoWrap>
+          {input.base ? (
+            <PreviewNumber value={input.base} />
+          ) : (
+            <TextInput data-key="base" value={state.base} onChange={onChange} />
+          )}
+        </Handle>
+
+        <Handle id="ratio">
+          <LabelNoWrap>Ratio</LabelNoWrap>
+          {input.ratio ? (
+            <PreviewNumber value={input.ratio} />
+          ) : (
+            <TextInput
+              data-key="ratio"
+              value={state.ratio}
+              onChange={onChange}
+            />
+          )}
+        </Handle>
+
+        <Handle id="precision">
+          <Stack direction="row" justify="between" align="center" gap={3}>
+            <Text>Precision</Text>
+            {input.precision !== undefined ? (
+              <Text>{input.precision}</Text>
+            ) : (
+              <TextInput
+                data-key="precision"
+                value={state.precision}
+                onChange={onChange}
+              />
+            )}
+          </Stack>
         </Handle>
       </HandleContainer>
       <HandleContainer type="source">