+++ packages/graph-engine/src/types.ts
@@ -84,6 +84,7 @@ export enum NodeTypes {
 
   CSS_MAP = "studio.tokens.css.map",
   CSS_BOX = "studio.tokens.css.box",
+  CSS_FUNCTIONS = "studio.tokens.css.function",
 
   //Logic
   IF = "studio.tokens.logic.if",
@@ -102,6 +103,8 @@ export enum NodeTypes {
   JOIN = "studio.tokens.array.join",
   CONCAT = "studio.tokens.array.concat",
   DOT_PROP = "studio.tokens.array.dotProp",
+  ARRAY_PASS_UNIT = "studio.tokens.array.passUnit",
+  NAME = "studio.tokens.array.name",
 
   // Math
   ADD = "studio.tokens.math.add",
@@ -136,6 +139,7 @@ export enum NodeTypes {
   COLOR_DISTANCE = "studio.tokens.color.distance",
   COLOR_NAME = "studio.tokens.color.name",
   NEAREST_TOKENS = "studio.tokens.color.nearestTokens",
+  SET_COLOR_LUMINANCE = "studio.tokens.color.setColorLuminance",
 
   //Sets
   FLATTEN = "studio.tokens.sets.flatten",
@@ -162,6 +166,7 @@ export enum NodeTypes {
   REGEX = "studio.tokens.string.regex",
   PASS_UNIT = "studio.tokens.typing.passUnit",
   PARSE_UNIT = "studio.tokens.typing.parseUnit",
+  STRINGIFY = "studio.tokens.string.stringify",
 
   //Accessibility
   CONTRAST = "studio.tokens.accessibility.contrast",