+++ packages/ui/src/pages/index.tsx
@@ -1,10 +1,13 @@
 import { Box } from '@tokens-studio/ui';
 import { LiveProvider } from 'react-live';
-import { code, scope } from '#/components/preview/scope.tsx';
+import { scope } from '#/components/preview/scope.tsx';
 import { useDispatch } from '#/hooks/index.ts';
 import { useSelector } from 'react-redux';
 import React, { useState } from 'react';
-import { showJourneySelector } from '#/redux/selectors/index.ts';
+import {
+  previewCodeSelector,
+  showJourneySelector,
+} from '#/redux/selectors/index.ts';
 import Joyride, { CallBackProps, STATUS } from 'react-joyride';
 // @ts-ignore
 import { themes } from 'prism-react-renderer';
@@ -13,13 +16,12 @@ import { useTheme } from '#/hooks/useTheme.tsx';
 import { useJourney } from '#/journeys/basic.tsx';
 import { JoyrideTooltip } from '#/components/joyride/tooltip.tsx';
 import { EditorTab } from '#/components/editor/index.tsx';
-import { PreviewContextProvider } from '#/providers/preview.tsx';
 
 const Wrapper = () => {
-  const [theCode, setTheCode] = useState(code);
   const dispatch = useDispatch();
   const showJourney = useSelector(showJourneySelector);
   const theme = useTheme();
+  const previewCode = useSelector(previewCodeSelector);
 
   const [{ steps }] = useJourney();
   const handleJoyrideCallback = (data: CallBackProps) => {
@@ -59,20 +61,19 @@ const Wrapper = () => {
           height: '100%',
           overflow: 'hidden',
           background: '$bgDefault',
+          isolation: 'isolate',
         }}
       >
-        <PreviewContextProvider code={theCode} setCode={setTheCode}>
-          <LiveProvider
-            code={theCode}
-            scope={scope}
-            theme={theme === 'light' ? themes.vsLight : themes.vsDark}
-            noInline={true}
-            enableTypeScript={true}
-            language="jsx"
-          >
-            <EditorTab setTheCode={setTheCode} />
-          </LiveProvider>
-        </PreviewContextProvider>
+        <LiveProvider
+          code={previewCode}
+          scope={scope}
+          theme={theme === 'light' ? themes.vsLight : themes.vsDark}
+          noInline={true}
+          enableTypeScript={true}
+          language="jsx"
+        >
+          <EditorTab />
+        </LiveProvider>
       </Box>
     </>
   );